+++ HISTORY.rst
@@ -2,6 +2,19 @@
 History
 =======
 
+0.3.0 (2023-10-17)
+------------------
+
+**Features and Improvements**
+
+* Revise :func:`set_expired` to work in the case where a response from
+  :func:`cert_please` or :func:`certs_please` was serialized - e.g. via :func:`json.dumps` -
+  and then de-serialized - e.g. via :func:`json.loads`.
+* Add :meth:`CertHero.from_dict` to create a :class:`CertHero` from a ``dict`` object
+  (unused for now).
+* Add integration tests to confirm the intended (correct) functionality.
+* Add a chart diagram for unzipped vs. zipped size comparisons between ``pyOpenSSL`` and ``asn1crypto``.
+
 0.2.0 (2023-10-17)
 ------------------